Small proofreading edits
In addition to the edits made directly within the text, please note the following small problems:
Line 6 - I think this sentence needs to be restructured. For example, I'm not sure what "two" refers to.
Line 167 - It looks like there is part of a sentence missing: there is the word "unreProviding" which seems to be the end of unfinished sentence running into a new sentence.